Namespace hive device table
