+++ app/models/device_orama_hive.rb
@@ -1,1 +1,4 @@
 require 'device_orama_hive/engine'
+
+module DeviceOramaHive
+end
