--- device_orama_hive/app/models/attribute.rb
@@ -1,5 +1,0 @@
-class Attribute < ActiveRecord::Base
-  def name
-    self.hostname
-  end
-end
