+++ device_orama_hive/app/models/device_orama_hive/attribute.rb
@@ -1,0 +1,7 @@
+module DeviceOramaHive
+  class Attribute < ActiveRecord::Base
+    def name
+      self.hostname
+    end
+  end
+end
