+++ device_orama_hive/db/migrate/20151110102253_create_device_orama_hive_attributes.rb
@@ -1,6 +1,6 @@
-class CreateAttributes < ActiveRecord::Migration
+class CreateDeviceOramaHiveAttributes < ActiveRecord::Migration
   def change
-    create_table :attributes do |t|
+    create_table :device_orama_hive_attributes do |t|
       t.string :hostname
       t.string :ip
       t.string :mac
