+++ device_orama_hive/spec/models/device_orama_hive/attribute_spec.rb
@@ -1,0 +1,15 @@
+require 'rails_helper'
+
+module DeviceOramaHive
+  RSpec.describe Attribute, type: :model do
+    describe '#name' do
+      let(:hive_attributes) {
+        Attribute.create( hostname: 'hive_host_name' )
+      }
+
+      it 'returns the hostname' do
+        expect(hive_attributes.name).to eq 'hive_host_name'
+      end
+    end
+  end
+end
